requirements.txt
Pillow
PyQt6
PyQt6_sip